LPD-87027 Rename and move
