+++ modules/apps/site/site-impl/src/main/java/com/liferay/site/internal/util/SitesImpl.java
@@ -746,7 +746,7 @@ public class SitesImpl implements Sites {
 						importLayoutSetPrototype(
 							layoutSetPrototype, groupId, privateLayout,
 							_layoutSetPrototypeHelper.
-								getLayoutSetPrototypesParameters(true),
+								getLayoutSetPrototypeParameters(true),
 							true);
 					}
 					finally {
