+++ portal-impl/src/com/liferay/portlet/exportimport/service/impl/ExportImportLocalServiceImpl.java
@@ -639,7 +639,7 @@ public class ExportImportLocalServiceImpl
 			BackgroundTaskManagerUtil.addBackgroundTask(
 				userId, groupId, exportImportConfiguration.getName(),
 				BackgroundTaskExecutorNames.
-					LAYOUT_SET_PROTOTYPE_SYNC_BACKGROUND_TASK_EXECUTOR,
+					LAYOUT_SET_PROTOTYPE_SYNC_EXPORT_IMPORT_BACKGROUND_TASK_EXECUTOR,
 				HashMapBuilder.<String, Serializable>put(
 					BackgroundTaskContextMapConstants.SKIP_NOTIFICATION, true
 				).put(
